Changed Column Type

+++ database/migrations/2025_03_18_095559_create_madeline_sessions_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('madeline_sessions', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id')->unique();
+            $table->unsignedBigInteger('user_id')->unique();
             $table->string('session_id')->unique();
             $table->timestamps();
         });
